requirements.txt
crewai
openai==0.28.1
streamlit==1.39.0
langchain==0.0.351
python-docx
